[#3545] minor folder restructure
