+++ akvo/rsr/spa/app/modules/editor/descriptions/action-types.js
