+++ akvo/rsr/spa/app/modules/editor/descriptions/actions.js
