+++ akvo/rsr/spa/app/modules/editor/descriptions/reducers.js
