+++ akvo/rsr/spa/app/modules/editor/descriptions/styles.scss
