+++ akvo/rsr/spa/app/modules/editor/info/info.js
