+++ akvo/rsr/spa/app/modules/editor/info/styles.scss
