+++ akvo/rsr/spa/app/modules/editor/partners/action-types.js
