+++ akvo/rsr/spa/app/modules/editor/partners/actions.js
