+++ akvo/rsr/spa/app/modules/editor/partners/partners.js
