+++ akvo/rsr/spa/app/modules/editor/partners/reducers.js
