+++ akvo/rsr/spa/app/modules/editor/partners/styles.scss
