+++ akvo/rsr/spa/app/modules/editor/settings/settings.js
