+++ akvo/rsr/spa/app/modules/editor/settings/styles.scss
