+++ akvo/rsr/spa/app/store/root-reducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux'
 
-import partnersRdr from '../modules/editor/modules/partners/reducers'
-import descsRdr from '../modules/editor/modules/descriptions/reducers'
+import partnersRdr from '../modules/editor/partners/reducers'
+import descsRdr from '../modules/editor/descriptions/reducers'
 
 const rootReducer = combineReducers({
   partnersRdr,
